requirements.txt
Flask==3.0.0
xlsxwriter==3.1.9
APScheduler==3.10.4
python-telegram-bot==20.7
pyinstaller==6.17.0
openpyxl==3.1.2
Flask-Login==0.6.3
Pillow
